Added the qr codes, blockchain addresses, converted them and others to links that will open their address in a blockchain explorer.
Also, adjusted the navigation so that the shop and contribute buttons were down low with the about button, and the shop tab now is just a link to the shop page instead of loading in the application.  The contribute button (Bitcoin logo) now just opens the about page with the contribute information.  Did this to lower the resources required by the application for mulitple browser instances.